moves Navbar to outer scope and use in default and error layouts
